move print trace to error.hpp
